Removed username requirement.
Previously, your username was required on the command line in order to generate
the client access token. But as it turns out, it's actually to name a file
caching this token. So it can be held static, and when a username is used later
in the program, the logged in user can be retrieved from the API.